uint64

+++ x/confidentialtransfers/keeper/keeper.go
@@ -29,7 +29,7 @@ type Keeper interface {
 	SetParams(ctx sdk.Context, params types.Params)
 
 	IsCtModuleEnabled(ctx sdk.Context) bool
-	GetRangeProofGasCost(ctx sdk.Context) uint32
+	GetRangeProofGasCost(ctx sdk.Context) uint64
 
 	BankKeeper() types.BankKeeper
 
@@ -186,8 +186,8 @@ func (k BaseKeeper) IsCtModuleEnabled(ctx sdk.Context) bool {
 }
 
 // GetRangeProofGasCost retrieves the value of the RangeProofGasCost param from the parameter store
-func (k BaseKeeper) GetRangeProofGasCost(ctx sdk.Context) uint32 {
-	var rangeProofGas uint32
+func (k BaseKeeper) GetRangeProofGasCost(ctx sdk.Context) uint64 {
+	var rangeProofGas uint64
 	k.paramSpace.Get(ctx, types.KeyRangeProofGas, &rangeProofGas)
 	return rangeProofGas
 }
